chore(env): create datadog-agent-commons crate for Agent-specific shared code
