+++ bin/agent-data-plane/src/components/apm_onboarding/install_info.rs
@@ -1,12 +1,11 @@
 use std::io::ErrorKind;
 
+use datadog_agent_commons::platform::PlatformSettings;
 use saluki_common::time::get_unix_timestamp;
 use saluki_error::{ErrorContext as _, GenericError};
 use serde::{Deserialize, Serialize};
 use stringtheory::MetaString;
 use uuid::Uuid;
-
-use crate::internal::platform::PlatformSettings;
 
 static INSTALL_TYPE_DEFAULT: MetaString = MetaString::from_static("manual");
 static INSTALL_TYPE_DOCKER_DEFAULT: MetaString = MetaString::from_static("docker_manual");
