+++ bin/agent-data-plane/src/components/mod.rs
@@ -2,6 +2,7 @@ pub mod apm_onboarding;
 mod dogstatsd_filterlist;
 pub mod dogstatsd_post_aggregate_filter;
 pub mod dogstatsd_prefix_filter;
+pub mod host_tags;
 pub mod ottl_filter_processor;
 pub mod ottl_transform_processor;
 pub mod tag_filterlist;
