+++ lib/datadog-agent-commons/src/ipc/client/bearer_auth.rs
