+++ lib/datadog-agent-commons/src/ipc/mod.rs
@@ -1,0 +1,6 @@
+//! IPC primitives for interacting with the Datadog Agent.
+
+pub mod client;
+pub mod config;
+pub mod session;
+pub mod tls;
