+++ lib/datadog-agent-commons/src/platform/macos_impl.rs
