--- lib/saluki-env/src/helpers/mod.rs
@@ -1,4 +1,0 @@
-//! Miscellaneous helpers used by different providers.
-
-pub mod remote_agent;
-pub mod tonic;
