--- lib/saluki-env/src/helpers/remote_agent/mod.rs
@@ -1,7 +1,0 @@
-//! Helpers for interacting with the Datadog Agent.
-
-mod client;
-pub use self::client::RemoteAgentClient;
-
-mod session;
-pub use self::session::{SessionId, SessionIdHandle};
