+++ lib/saluki-env/src/host/providers/mod.rs
@@ -5,6 +5,3 @@ pub use self::boxed::BoxedHostProvider;
 
 mod fixed;
 pub use self::fixed::FixedHostProvider;
-
-mod remote_agent;
-pub use self::remote_agent::RemoteAgentHostProvider;
