+++ lib/saluki-env/src/workload/providers/mod.rs
@@ -2,6 +2,3 @@
 
 mod noop;
 pub use self::noop::NoopWorkloadProvider;
-
-mod remote_agent;
-pub use self::remote_agent::{RemoteAgentWorkloadAPIHandler, RemoteAgentWorkloadProvider};
