+++ lib/saluki-io/src/net/mod.rs
@@ -13,10 +13,3 @@ pub use self::stream::{Connection, Stream};
 pub mod unix;
 
 pub mod util;
-
-mod ipc;
-pub use rustls::ServerConfig;
-
-pub use self::ipc::{
-    build_datadog_agent_client_ipc_tls_config, build_datadog_agent_server_tls_config, get_ipc_cert_file_path,
-};
